requirements.txt
requests>=2.31,<3.0
paho-mqtt>=2.1.0,<3.0
pillow>=10.3,<11.0
pyaml>=23.12,<24.0
pydantic>=2.6,<3.0
pysocks>=1.7,<2.0
protobuf>=3.20,<4.0
certifi>=2024.2
deprecated>=1.2,<2.0
rdp2==1.1.1
opentelemetry-api==1.36.0
opentelemetry-sdk==1.36.0
